enhance UI to reduce user from deleting entry too easily (#3)

+++ Eklee.ActivityTracker/Pages/Home.razor.cs
@@ -15,6 +15,8 @@ public partial class Home
     private ActivitySession? activitySession;
     private HomeActivity? activeActivity;
     private RadzenDataList<HomeActivity>? radzenDataList;
+    public bool SelectMode { get; set; }
+    private DateTime Current { get; set; }
 
     protected override async Task OnInitializedAsync()
     {
@@ -25,16 +27,17 @@ public partial class Home
     {
         activities.Clear();
         activities.AddRange([.. (await ActivityService!.GetActivitiesAsync()).Select(x => new HomeActivity(x)).OrderByDescending(x => x.GetActivity().LastUpdated)]);
-        await radzenDataList!.FirstPage();        
+        await radzenDataList!.FirstPage();
     }
 
     private void SelectActivity(HomeActivity activity)
     {
+        Current = DateTime.Now;
         if (activitySession is null)
         {
             activitySession = new ActivitySession
             {
-                Start = DateTime.Now
+                Start = Current
             };
 
             activitySession.Items ??= [];
@@ -42,10 +45,21 @@ public partial class Home
 
         var item = new ActivityItem
         {
-            Start = DateTime.Now
+            Start = Current,
         };
         activitySession.Items = [.. activitySession.Items, item];
         activeActivity = activity;
+    }
+
+    private void RestartActivity()
+    {
+        if (activitySession is null || activeActivity is null)
+        {
+            return;
+        }
+
+        activitySession.Items![^1].Start = DateTime.Now;
+        Current = activitySession.Items![^1].Start!.Value;
     }
 
     private async Task StopActivityAsync()
@@ -89,9 +103,10 @@ public partial class Home
 
         activitySession.Items![^1].DurationInSeconds = Convert.ToInt32((DateTime.Now - activitySession.Items![^1].Start!.Value).TotalSeconds);
 
+        Current = DateTime.Now;
         activitySession.Items = [.. activitySession.Items, new ActivityItem
         {
-            Start = DateTime.Now
+            Start = Current
         }];
     }
 
@@ -112,9 +127,32 @@ public partial class Home
         }
     }
 
-    private async Task DeleteActivity(HomeActivity activity)
+    private async Task DeleteActivity()
     {
-        await ActivityService!.DeleteActivityAsync(activity.GetActivity());
+        var toDelete = activities.Where(x => x.Selected);
+        if (!toDelete.Any())
+        {
+            return;
+        }
+
+        foreach (var activity in toDelete)
+        {
+            await ActivityService!.DeleteActivityAsync(activity.GetActivity());
+        }
         await RefreshAsync();
+    }
+
+    private void EnableSelectMode()
+    {
+        SelectMode = true;
+    }
+
+    private void DisableSelectMode()
+    {
+        foreach (var activity in activities.Where(x => x.Selected))
+        {
+            activity.Selected = false;
+        }
+        SelectMode = false;
     }
 }
